refactor useKeypress from customhooks to utils/hooks.ts
