+++ src/renderer/components/NewProjectModal/ImportMediaView.tsx
@@ -4,7 +4,7 @@ import { Box, Stack, styled, Typography } from '@mui/material';
 import colors from 'renderer/colors';
 import CloseIcon from '@mui/icons-material/Close';
 import IconButton from '@mui/material/IconButton';
-import useKeypress from 'renderer/hooks/customHooks';
+import useKeypress from 'renderer/utils/hooks';
 import { ApplicationStore } from '../../store/sharedHelpers';
 import { projectCreated } from '../../store/currentProject/actions';
 import { RuntimeProject } from '../../../sharedTypes';
