+++ src/renderer/components/NewProjectModal/NewProjectView.tsx
@@ -3,7 +3,7 @@ import CloseIcon from '@mui/icons-material/Close';
 import IconButton from '@mui/material/IconButton';
 import { ChangeEvent, useCallback, useState } from 'react';
 import { useDispatch } from 'react-redux';
-import useKeypress from 'renderer/hooks/customHooks';
+import useKeypress from 'renderer/utils/hooks';
 import { makeProjectWithoutMedia } from '../../util';
 import { projectCreated } from '../../store/currentProject/actions';
 import colors from '../../colors';
